requirements.txt
streamlit
opencv-python
cvzone
tensorflow
numpy
pillow
